requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
dateparser==1.2.0
demjson3==3.0.6
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
facebook-scraper==0.2.59
fake-useragent==1.3.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
lxml==4.9.3
mutagen==1.47.0
oauthlib==3.2.2
outcome==1.3.0.post0
parse==1.19.1
playwright==1.40.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pyee==11.0.1
pyparsing==3.1.1
pyppeteer==0.0.25
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
TikTokApi==6.2.0
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
wsproto==1.2.0
yt-dlp==2023.11.16
zipp==3.17.0
